Adds powersports api
