+++ tests/PowersportsTest.php
@@ -3,18 +3,18 @@ declare(strict_types=1);
 
 namespace Test;
 
-use CarApiSdk\CarApi;
 use CarApiSdk\CarApiConfig;
 use CarApiSdk\CarApiException;
 use CarApiSdk\JsonSearch;
 use CarApiSdk\JsonSearchItem;
+use CarApiSdk\Powersports;
 use Http\Discovery\Psr17FactoryDiscovery;
 use Http\Discovery\Psr18Client;
 use Nyholm\Psr7\Response;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-class CarApiTest extends TestCase
+class PowersportsTest extends TestCase
 {
     /**
      * @dataProvider dataProviderForMethods
@@ -23,7 +23,7 @@ class CarApiTest extends TestCase
     {
         $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
         $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
+        $sdk = new Powersports($config, $client);
         $obj = $sdk->{$method}();
         $this->assertObjectHasProperty('data', $obj);
     }
@@ -33,16 +33,6 @@ class CarApiTest extends TestCase
         return [
             ['makes'],
             ['models'],
-            ['trims'],
-            ['bodies'],
-            ['mileages'],
-            ['engines'],
-            ['interiorColors'],
-            ['exteriorColors'],
-            ['accountRequests'],
-            ['accountRequestsToday'],
-            ['csvDataFeedLastUpdated'],
-            ['obdCodes'],
         ];
     }
 
@@ -50,62 +40,9 @@ class CarApiTest extends TestCase
     {
         $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
         $client = $this->createMockClient(200, '["data"]');
-        $sdk = new CarApi($config, $client);
+        $sdk = new Powersports($config, $client);
         $arr = $sdk->years();
         $this->assertNotEmpty($arr);
-    }
-
-    public function test_trim_item(): void
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
-        $obj = $sdk->trimItem(1);
-        $this->assertObjectHasProperty('data', $obj);
-    }
-
-    public function test_vin(): void
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
-        $obj = $sdk->vin('123');
-        $this->assertObjectHasProperty('data', $obj);
-    }
-
-    public function test_license_plate(): void
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
-        $obj = $sdk->licensePlate('US', 'LNP8460#TEST', 'NY');
-        $this->assertObjectHasProperty('data', $obj);
-    }
-
-    public function test_single_obd_code(): void
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
-        $obj = $sdk->obdCodeItem('B1200');
-        $this->assertObjectHasProperty('data', $obj);
-    }
-
-    public function test_vehicle_attributes(): void
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '["data"]');
-        $sdk = new CarApi($config, $client);
-        $arr = $sdk->vehicleAttributes('tesst');
-        $this->assertNotEmpty($arr);
-    }
-
-    public function test_csv_datafeed()
-    {
-        $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
-        $client = $this->createMockClient(200, '');
-        $sdk = new CarApi($config, $client);
-        $this->assertInstanceOf(\Psr\Http\Message\ResponseInterface::class, $sdk->csvDataFeed());
     }
 
     public function test_exception_response(): void
@@ -119,7 +56,7 @@ class CarApiTest extends TestCase
           "url": "/url/path",
           "message": "Internal Error"
         }');
-        $sdk = new CarApi($config, $client);
+        $sdk = new Powersports($config, $client);
 
         $this->expectException(CarApiException::class);
         $this->expectExceptionMessage('ExceptionName: Internal Error while requesting /url/path');
@@ -130,7 +67,7 @@ class CarApiTest extends TestCase
     {
         $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
         $client = $this->createMockClient(200, 'bad json');
-        $sdk = new CarApi($config, $client);
+        $sdk = new Powersports($config, $client);
 
         $this->expectException(CarApiException::class);
         $this->expectExceptionMessage('Error decoding response');
@@ -141,7 +78,7 @@ class CarApiTest extends TestCase
     {
         $config = CarApiConfig::build(['token' => '1', 'secret' => '1']);
         $client = $this->createMockClient(200, '{"data": []}');
-        $sdk = new CarApi($config, $client);
+        $sdk = new Powersports($config, $client);
 
         $json = (new JsonSearch())
             ->addItem(new JsonSearchItem('make', 'in', ['Tesla']));
